requirements.txt
# Data Collection
requests==2.31.0
tweepy==4.14.0
praw==7.7.1
python-dotenv==1.0.0

# Data Processing
pandas==2.0.3
numpy==1.24.3
openpyxl==3.1.2
python-dateutil==2.8.2

# NLP & Text Processing
spacy==3.6.0
textblob==0.17.1
nltk==3.8.1
wordcloud==1.9.2

# Machine Learning & Topic Modeling
scikit-learn==1.3.0
xgboost==1.7.6
lightgbm==4.0.0
sentence-transformers==2.2.2
gensim==4.3.1

# Time Series Analysis
prophet==1.1.5
statsmodels==0.14.0

# Visualization
matplotlib==3.7.2
seaborn==0.12.2
plotly==5.15.0
wordcloud==1.9.2

# Web Application
streamlit==1.25.0

# Utilities
tqdm==4.65.0
python-dotenv==1.0.0
pytest==7.4.0
